fix: call OWALChanges.setShortValue() when setting a short value
This affected big endian platforms. The test CellBTreeMultiValueV2TestIT
fails on big endian platforms without this fix.

+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/base/ODurablePage.java
@@ -261,7 +261,7 @@ public class ODurablePage {
   protected final int setShortValue(final int pageOffset, final short value) {
 
     if (changes != null) {
-      changes.setIntValue(buffer, value, pageOffset);
+      changes.setShortValue(buffer, value, pageOffset);
     } else {
       assert buffer != null;
       assert buffer.order() == ByteOrder.nativeOrder();
